Finished debugging the power system
Power system now works with a discharging battery. The models for a DC machine using the Claytex library are also included, but they are incomplete.

The battery model requires the Modelon Library Suite to run, so I also included a model of a ramp voltage to act as a battery for those who don't use the library.