Update SimpleDCmotor units
Updated units according to: https://lup.lub.lu.se/luur/download?func=downloadFile&recordOId=8847641&fileOId=8859343

Because we are using simplified calculations , I created a new variable type thrust and drag to correct the units for those two coefficients. The reason they are approximated in the library is that they are commonly approximated in papers and other calculations as they are dependent on factors with limited variability such as air density.